+++ Team-3749-2025/src/main/java/frc/robot/subsystems/elevator/sim/ElevatorSimulation.java
@@ -1,8 +1,12 @@
-package frc.robot.subsystems.elevator;
+package frc.robot.subsystems.elevator.sim;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.wpilibj.simulation.ElevatorSim;
+import frc.robot.subsystems.elevator.ElevatorConstants;
+import frc.robot.subsystems.elevator.ElevatorIO;
+import frc.robot.subsystems.elevator.ElevatorConstants.ElevatorSpecs;
+import frc.robot.subsystems.elevator.ElevatorIO.ElevatorData;
 import frc.robot.utils.MiscConstants.SimConstants;
 
 /**
@@ -27,7 +31,7 @@ public class ElevatorSimulation implements ElevatorIO {
         ElevatorConstants.ElevatorSpecs.startingHeightMeters
     );
 
-    ElevatorSimulation (){
+    public ElevatorSimulation (){
         System.out.println("[Init] Creating ElevatorSimulation");
     }
